Module google_verifiedaccess1::api

Structs

Challenge
Result message for VerifiedAccess.CreateChallenge.

ChallengeCreateCall
CreateChallenge API

ChallengeMethods
A builder providing access to all methods supported on challenge resources. It is not used directly, but through the Verifiedaccess hub.

ChallengeVerifyCall
VerifyChallengeResponse API

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

SignedData
The wrapper message of any data and its signature.

Verifiedaccess
Central instance to access all Verifiedaccess related resource activities

VerifyChallengeResponseRequest
signed ChallengeResponse

VerifyChallengeResponseResult
Result message for VerifiedAccess.VerifyChallengeResponse.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.